+++ daemons/common/avtp.c
@@ -33,12 +33,10 @@
 #include "pdu.h"
 #include "avtp.h"
 
-ssize_t avtp_common_control_header_read(struct avtp_common_control_header *p,
-					void const *base, ssize_t pos,
-					size_t len)
+int avtp_common_control_header_read(struct avtp_common_control_header *p,
+				    void const *base, int pos, int len)
 {
-	ssize_t r =
-	    pdu_validate_range(pos, len, AVTP_COMMON_CONTROL_HEADER_LEN);
+	int r = pdu_validate_range(pos, len, AVTP_COMMON_CONTROL_HEADER_LEN);
 	if (r >= 0) {
 		p->subtype = avtp_common_control_header_get_subtype(base, pos);
 		p->sv = avtp_common_control_header_get_sv(base, pos);
@@ -55,12 +53,10 @@ ssize_t avtp_common_control_header_read(struct avtp_common_control_header *p,
 	return r;
 }
 
-ssize_t
-avtp_common_control_header_write(struct avtp_common_control_header const *p,
-				 void *base, ssize_t pos, size_t len)
+int avtp_common_control_header_write(struct avtp_common_control_header const *p,
+				     void *base, int pos, int len)
 {
-	ssize_t r =
-	    pdu_validate_range(pos, len, AVTP_COMMON_CONTROL_HEADER_LEN);
+	int r = pdu_validate_range(pos, len, AVTP_COMMON_CONTROL_HEADER_LEN);
 	if (r >= 0) {
 		avtp_common_control_header_set_subtype(p->subtype, base, pos);
 		avtp_common_control_header_set_sv(p->sv, base, pos);
